+++ lib/scss_lint/linter/indentation.rb
@@ -49,7 +49,7 @@ module SCSSLint
         return true
       end
 
-      unless actual_indent.length == @indent
+      unless allow_arbitrary_indent?(node) || actual_indent.length == @indent
         add_lint(node.line,
                  "Line should be indented #{@indent} #{character_name}s, " \
                  "but was indented #{actual_indent.length} #{character_name}s")
@@ -129,6 +129,12 @@ module SCSSLint
       return unless first_child_source = node.children.first.source_range
 
       same_position?(node.source_range.end_pos, first_child_source.start_pos)
+    end
+
+    def allow_arbitrary_indent?(node)
+      @indent == 0 &&
+        config['allow_non_nested_indentation'] &&
+        node.is_a?(Sass::Tree::RuleNode)
     end
   end
 end